add speed to interface and refine fdtclient

+++ miscellaneous/fdtclient/fdtclient.py
@@ -3,7 +3,14 @@ import fileinput
 import re
 import Pyro4
 import sys
+import os
 
+
+
+"""
+all qos for an interface should be specified in the same conf file
+different interfaces can use different conf file
+"""
 class FdtClient:
 
     def __init__(self, jobId, remoteHost, fileName, fdtJarLocation, interface, remotePort):
@@ -13,7 +20,7 @@ class FdtClient:
         self.fdtJarLocation = fdtJarLocation
         self.fileName = fileName
         self.interface = interface
-        self.qosfilename = 'fireqos.conf'
+        self.qosfilename = "./fireqos-" + str(interface) +".conf" # should be shared with different clients
 
         self.clientPorts = []
 
@@ -23,10 +30,12 @@ class FdtClient:
     #   class job1 commit 6mbit max 8mbit
     #      match4 src 172.27.239.196
     def __generateQosConfFile(self, interface, speed):
-        f = open(self.qosfilename, 'w')
-        f.write("interface " + str(interface) + " world-in input rate " + str(speed))
-        f.flush()
-        f.close()
+        if not os.path.exists(self.qosfilename):
+            f = open(self.qosfilename, 'w')
+            f.write("interface " + str(interface) + " world-in-" + str(interface) + " input rate " + str(speed))
+            f.flush()
+            f.close()
+
 
 
     #classname = jobID + clientport
@@ -57,15 +66,24 @@ class FdtClient:
 
 
 
-    #java -jar fdt.jar -c qiao-thinkstation -pull -d ./http.log.fdt ./http.log
-    def startClient(self):
-        p = Popen(['java', '-jar', self.fdtJarLocation + 'fdt.jar', '-c', self.remoteHost, '-pull', '-d',
-                   self.fdtJarLocation + str(self.jobId), './' + self.fileName])
+    #java -jar fdt.jar -c 10.10.14.6 -P 33 -wCount 100 -pull -d /dev/null /dev/zero
+    def startClient(self, interfaceSpeed):
+        p = Popen(['java', '-jar', self.fdtJarLocation, '-c', self.remoteHost, '-P', '33', '-wCount', '100', '-pull', '-d',
+                   self.fileName, '/dev/null'])
 
-        self.__setCLientPorts()
+        self.__setClientPorts()
 
-        self.__generateQosConfFile(self.interface)
+        self.__generateQosConfFile(self.interface, interfaceSpeed)
         print("started fdt")
+
+
+    def __getControlPort(self):
+        min = 99999
+        for port in self.clientPorts:
+            if int(port) < min:
+                min = int(port)
+        return min
+
 
     #tcp6       0      0 [UNKNOWN]:57052         qn-in-xbd.1e100.n:https ESTABLISHED -
     def __setClientPorts(self):
@@ -85,14 +103,19 @@ class FdtClient:
 
 
     def changeRate(self, newRate):
-        eachRate = newRate / len(self.clientPorts)
+        eachRate = newRate / (len(self.clientPorts) - 1)
 
         for port in self.clientPorts:
+            if int(port) == self.__getControlPort():
+                continue
             self.__changeRateForClass(port, eachRate)
 
+        self.__finishChangeRate()
 
-    def getReceivedDataSize(self):
-        return
+
+    def __finishChangeRate(self):
+        p = Popen(['fireqos', 'start', self.qosfilename])
+
 
 @Pyro4.expose
 class FdtClientManager:
@@ -100,24 +123,28 @@ class FdtClientManager:
     def __init__(self):
         self.jobId2fdtClient = {}
 
-        self.ip2Interface = {}
+        self.ip2Interface2Speed = {}
         self.__initializeIp2Interface()
 
 
     def __initializeIp2Interface(self):
-        with open("./ip2interface", "r") as ip2interfaceFile:
+        with open("./ip2interface2rate", "r") as ip2interfaceFile:
             for line in ip2interfaceFile:
                 ip = str(line).split(" ")[0]
                 interface = str(line).split(" ")[1]
-                self.ip2Interface[ip] = interface
+                speed = str(line).split(" ")[2]
+                self.ip2Interface2Speed[ip] = interface + " " + speed
 
 
+    # rate is kbit
     def startDataTransferWithRate(self, jobId, remoteHost, localHost, fileName, rate):
         fdtJarLocation = "./fdt.jar"
-        interface = self.ip2Interface[localHost]
+        interface = str(self.ip2Interface2Speed[localHost]).split(" ")[0]
+        speed = str(self.ip2Interface2Speed[localHost]).split(" ")[1]
         remotePort = 54321
 
         fdtClient = self.__getAFdtClient(jobId, remoteHost, fileName, fdtJarLocation, interface, remotePort)
+        fdtClient.startClient(speed)
         fdtClient.changeRate(rate)
